fix: support bf16 scales in GPTQ Marlin MoE

+++ python/sglang/srt/layers/quantization/gptq/schemes/gptq_moe.py
@@ -236,7 +236,9 @@ class GPTQMarlinMoEScheme(GPTQMoESchemeBase):
                 num_experts,
                 scales_size13,
                 2 * intermediate_size_per_partition,
-                dtype=torch.half,
+                # Marlin specializes scale loads for the activation dtype.
+                # Keep them aligned for both FP16 and BF16 model execution.
+                dtype=params_dtype,
             ),
             requires_grad=False,
         )
@@ -244,7 +246,9 @@ class GPTQMarlinMoEScheme(GPTQMoESchemeBase):
         set_weight_attrs(w13_scales, extra_weight_attrs)
 
         w2_scales = torch.nn.Parameter(
-            torch.empty(num_experts, scales_size2, hidden_size, dtype=torch.half),
+            torch.empty(
+                num_experts, scales_size2, hidden_size, dtype=params_dtype
+            ),
             requires_grad=False,
         )
         layer.register_parameter("w2_scales", w2_scales)
